Revert "Proste wykrywanie aktywności - baza dla dalszego rozwoju"
