+++ src/main/java/pl/edu/agh/io/umniedziala/MouseListener.java
@@ -1,0 +1,41 @@
+package pl.edu.agh.io.umniedziala;
+import org.jnativehook.GlobalScreen;
+import org.jnativehook.NativeHookException;
+import org.jnativehook.mouse.NativeMouseEvent;
+import org.jnativehook.mouse.NativeMouseInputListener;
+
+public class MouseListener implements NativeMouseInputListener{
+    public void nativeMouseClicked(NativeMouseEvent e) {
+        System.out.println("Mouse Clicked: " + e.getClickCount());
+    }
+
+    public void nativeMousePressed(NativeMouseEvent e) {
+        System.out.println("Mouse Pressed: " + e.getButton());
+    }
+
+    public void nativeMouseReleased(NativeMouseEvent e) {
+        System.out.println("Mouse Released: " + e.getButton());
+    }
+
+    public void nativeMouseMoved(NativeMouseEvent e) {
+        System.out.println("Mouse Moved: " + e.getX() + ", " + e.getY());
+    }
+
+    public void nativeMouseDragged(NativeMouseEvent e) {
+        System.out.println("Mouse Dragged: " + e.getX() + ", " + e.getY());
+    }
+
+    public static void main(String[] args) {
+        try {
+            GlobalScreen.registerNativeHook();
+        }
+        catch (NativeHookException ex) {
+            System.err.println("There was a problem registering the native hook.");
+            System.err.println(ex.getMessage());
+            System.exit(1);
+        }
+        MouseListener mouseListener = new MouseListener();
+        GlobalScreen.addNativeMouseListener(mouseListener);
+        GlobalScreen.addNativeMouseMotionListener(mouseListener);
+    }
+}
